+++ packages/user/src/model/invitations/controller.ts
@@ -7,11 +7,20 @@ const plugin = async (
   options: unknown,
   done: () => void
 ) => {
+  const ROUTE_INVITATIONS = "/invitations";
   const ROUTE_INVITATIONS_ACCEPT = "/invitations/token/:token";
   const ROUTE_INVITATIONS_CREATE = "/invitations";
   const ROUTE_INVITATIONS_GET_BY_TOKEN = "/invitations/token/:token";
   const ROUTE_INVITATIONS_RESEND = "/invitations/resend/:id(^\\d+)";
   const ROUTE_INVITATIONS_REVOKE = "/invitations/revoke/:id(^\\d+)";
+
+  fastify.get(
+    ROUTE_INVITATIONS,
+    {
+      preHandler: fastify.verifySession(),
+    },
+    handlers.listInvitation
+  );
 
   fastify.post(
     ROUTE_INVITATIONS_CREATE,