+++ packages/user/src/model/invitations/handlers/index.ts
@@ -1,6 +1,7 @@
 import acceptInvitation from "./acceptInvitation";
 import createInvitation from "./createInvitation";
 import getInvitationByToken from "./getInvitationByToken";
+import listInvitation from "./listInvitation";
 import resendInvitation from "./resendInvitation";
 import revokeInvitation from "./revokeInvitation";
 
@@ -8,6 +9,7 @@ export default {
   acceptInvitation,
   createInvitation,
   getInvitationByToken,
+  listInvitation,
   resendInvitation,
   revokeInvitation,
 };